backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.29.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.10.2
pydantic-settings==2.6.1
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.0
python-dotenv==1.0.1
httpx==0.27.0
faiss-cpu==1.8.0
sentence-transformers==2.2.2
langchain-huggingface==0.1.2
